Create ScraperWithRAUrls
Sorry, I am new to Python so my coding may not be the best.
I added the ability to send the ReserveAmerica URLs when sending the email. This will allow you book the site more quickly.
I also changed the site length to a greater than or equal, so you can put your rigs actual length.
Last thing is that I had to add some date formatting so that I could re-arrange the booking date format to comply with the RA URL requirement.